fix(dashboard): lint errors
