+++ dashboard/src/components/deployments/AppDeploymentDuration/index.ts
@@ -1,0 +1,2 @@
+export * from './AppDeploymentDuration';
+export { default } from './AppDeploymentDuration';
